pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tty-theme"
version = "0.1.0"
description = "AI terminal theme generator for Ghostty, iTerm2, and more"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Terminals",
]
dependencies = [
    "typer>=0.12",
    "pillow>=10.0",
    "scikit-learn>=1.4",
    "imagehash>=4.3",
    "sentence-transformers>=3.0",
    "keyring>=25.0",
    "httpx>=0.27",
    "numpy>=1.26",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "google-cloud-firestore>=2.16",
    "google-cloud-secret-manager>=2.20",
    "python-dotenv>=1.0",
    "prometheus-client>=0.20",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "bandit>=1.7",
    "pip-audit>=2.7",
    "python-dotenv>=1.0",
]

[project.scripts]
tty-theme = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["api", "cli", "modes", "generator", "image", "cache", "providers", "security", "themes", "schema"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "S", "B", "UP"]
ignore = [
    "S101",   # assert in tests is fine
    "S603",   # subprocess without shell=True is fine
    "S607",   # partial path for subprocess is fine
]
per-file-ignores = { "tests/*" = ["S101", "S106", "E501"], "cli/*" = ["B008"] }

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
